refactor: change link resolver in object event file
