--- packages/services/src/models/vckit.ts
@@ -1,9 +1,0 @@
-export const contextDefault = [
-    'https://www.w3.org/2018/credentials/v1',
-    'https://w3id.org/vc-revocation-list-2020/v1',
-    'https://w3id.org/security/suites/jws-2020/v1',
-    'https://dev-render-method-context.s3.ap-southeast-1.amazonaws.com/dev-render-method-context.json',
-    'https://w3id.org/security/suites/jws-2020/v1',
-];
-
-export const typeDefault = ['VerifiableCredential'];
